requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.1.0
tqdm>=4.66.0
python-dateutil>=2.8.2
tabulate>=0.9.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
